services/exporter/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
weasyprint
python-pptx
markdown
jinja2
sqlalchemy
psycopg2-binary
requests
